Rename original_registration to original_registration.py
This program asks student if the student is new or current student.
If the student is current student it asks for user name and password. 
If the user name and password is correct it prompts to add or remove or exit. But if the student is new it will create user name and prompts the student to create password.